+++ release/components/body/body.component.js
@@ -183,9 +183,8 @@ var DataTableBodyComponent = (function () {
      */
     DataTableBodyComponent.prototype.updateOffsetY = function (offset) {
         // scroller is missing on empty table
-        if (!this.scroller) {
+        if (!this.scroller)
             return;
-        }
         if (this.scrollbarV && offset) {
             // First get the row Index that we need to move to.
             var rowIndex = this.pageSize * offset;