+++ release/utils/sort.js
@@ -1,7 +1,7 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
 var types_1 = require("../types");
-var deep_getter_1 = require("./deep-getter");
+var column_prop_getters_1 = require("./column-prop-getters");
 /**
  * Gets the next sort direction
  * @param  {SortType}      sortType
@@ -83,15 +83,26 @@ function sortRows(rows, columns, dirs) {
         }
         return obj;
     }, {});
+    // cache valueGetter and compareFn so that they
+    // do not need to be looked-up in the sort function body
+    var cachedDirs = dirs.map(function (dir) {
+        var prop = dir.prop;
+        return {
+            prop: prop,
+            dir: dir.dir,
+            valueGetter: column_prop_getters_1.getterForProp(prop),
+            compareFn: cols[prop] || orderByComparator
+        };
+    });
     return temp.sort(function (a, b) {
-        for (var _i = 0, dirs_1 = dirs; _i < dirs_1.length; _i++) {
-            var _a = dirs_1[_i], prop = _a.prop, dir = _a.dir;
-            var propA = deep_getter_1.deepValueGetter(a, prop);
-            var propB = deep_getter_1.deepValueGetter(b, prop);
-            var compareFn = cols[prop] || orderByComparator;
-            var comparison = dir !== types_1.SortDirection.desc ?
-                compareFn(propA, propB) :
-                -compareFn(propA, propB);
+        for (var _i = 0, cachedDirs_1 = cachedDirs; _i < cachedDirs_1.length; _i++) {
+            var cachedDir = cachedDirs_1[_i];
+            var prop = cachedDir.prop, valueGetter = cachedDir.valueGetter;
+            var propA = valueGetter(a, prop);
+            var propB = valueGetter(b, prop);
+            var comparison = cachedDir.dir !== types_1.SortDirection.desc ?
+                cachedDir.compareFn(propA, propB) :
+                -cachedDir.compareFn(propA, propB);
             // Don't return 0 yet in case of needing to sort by next property
             if (comparison !== 0)
                 return comparison;